TX Numero arrumado
